# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets for current user
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Ticket:
      type: object
      required: [id, userId, subject, category, message, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        subject:
          type: string
        category:
          type: string
          description: Ticket category (e.g. "bug", "domanda", "altro").
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateTicketRequest:
      type: object
      required: [subject, category, message]
      properties:
        subject:
          type: string
          minLength: 1
          maxLength: 128
        category:
          type: string
          minLength: 1
          description: Ticket category (e.g. "bug", "domanda", "altro").
        message:
          type: string
          minLength: 1
...